Rename server and improve methods
Renamed server to communicator, added start(), stop(), and communicate().
